Simplify InputControl’s state reducer and event handling
The goal is to simplify InputControl by generalizing it while still
allowing the same extensibility. It is done by removing the dragging
and keyboard arrow/enter key handling along with their actions types
in the reducer. By passing the state reducer’s methods out through
the imperative handle hook, a component like NumberControl can add
the event handling hooked up to state dispatches as it needs instead
of relying only on what it can change within the reducer function.

This simplifies InputControl greatly while adding a bit of complexity
to NumberControl. Overall, it seems promising and tests well manually
but some strange test failures and others that require undesirable
edits to pass make it seem like it may not be the right approach.